backend/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
#readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10" 
anchorpy = "0.20.1"
solana = "^0.34.0"
requests = "^2.28.1"
#types-requests = "^2.28.9"
ccxt = "^4.0.0"
pandas = "^1.5.3"
pandas-ta = "^0.3.14b0"
anchorpy-core = "0.2.0"
python-dotenv = "^1.0.0"
#drift-jit-proxy = "0.1.4"
#asyncio = "^3.4.3"
driftpy = "0.7.89"
Backtesting = "^0.3.3"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"


[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
pytest-asyncio = "^0.21.0"
pytest-xprocess = "0.18.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
